refactor client
